+++ chai/defs.h
@@ -14,6 +14,7 @@ using namespace std;
 
 //#define TESTING
 //#define ASSERT
+#define INFO
 
 /**
  * Write logging information into log.txt file.